pyproject.toml
[project]
name = "photo-archivist"
version = "0.1.0"
description = "Intelligent photo archival and printing system"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn[standard]>=0.23.0",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
    "sqlmodel>=0.0.8",
    "aiosqlite>=0.19.0",
    "msal>=1.24.0",
    "httpx>=0.25.0",
    "opencv-python-headless>=4.8.0.76",
    "pillow>=10.0.0",
    "imagehash>=4.3.1",
    "APScheduler>=3.10.4",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    "pydantic[email]>=2.4.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-env>=1.0.0",
    "pytest-mock>=3.11.0",
    "ruff>=0.0.287",
    "black>=23.7.0",
    "mypy>=1.5.1",
    "types-pillow>=10.0.0",
    "types-python-dateutil>=2.8.19.14",
]

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]
ignore = ["E501"]  # Line length handled by black

[tool.ruff.isort]
known-first-party = ["app"]
force-single-line = true

[tool.black]
line-length = 100
target-version = ["py39"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
plugins = ["pydantic.mypy"]
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["imagehash.*", "cv2.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --cov=app --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning:pkg_resources.*",
]